- id изменил на long, потому что в телеграме они больше int уже;

+++ webservice/NotificationService/BotNotificationService/Models/Update.cs
@@ -7,7 +7,7 @@ namespace BotNotificationService.Models
 {
     public class Update
     {
-        public int update_id { get; set; }
+        public long update_id { get; set; }
         public Message message { get; set; }
     }
 }
